Moved run_query method into the Reporter module and generalized it for reports.  Have implemented it in JSR and ER.
